% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficients for ridgereg}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{a ridgereg object}

\item{...}{additional arguments (ignored)}
}
\value{
numeric vector of coefficients
}
\description{
Coefficients for ridgereg
}
