% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ridgereg.R
\name{pred.ridgereg}
\alias{pred.ridgereg}
\title{Predicted/fitted values for ridgereg}
\usage{
\method{pred}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a ridgereg object}

\item{newdata}{optional data.frame for prediction}

\item{...}{additional arguments (ignored)}
}
\value{
numeric vector of fitted or predicted values
}
\description{
Predicted/fitted values for ridgereg
}
