% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method for ridgereg}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a ridgereg object}

\item{newdata}{optional data.frame for prediction}

\item{...}{additional arguments (ignored)}
}
\value{
numeric vector of predicted values
}
\description{
Predict method for ridgereg
}
