% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print method for ridgereg}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{a ridgereg object}

\item{...}{additional arguments (ignored)}
}
\description{
Print method for ridgereg
}
