% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ridgereg.R
\name{summary.ridgereg}
\alias{summary.ridgereg}
\title{Summary method for ridgereg}
\usage{
\method{summary}{ridgereg}(object, ...)
}
\arguments{
\item{object}{a ridgereg object}

\item{...}{additional arguments (ignored)}
}
\description{
Summary method for ridgereg
}
